+++ python/GafferUI/PlugLayout.py
@@ -40,6 +40,8 @@ import sys
 import functools
 import collections
 
+import imath
+
 import IECore
 
 import Gaffer
@@ -140,7 +142,7 @@ class PlugLayout( GafferUI.Widget ) :
 		self.__contextTracker = GafferUI.ContextTracker.acquireForFocus( parent )
 		self.__contextTracker.changedSignal( parent ).connect( Gaffer.WeakMethod( self.__contextChanged ) )
 
-		self.__filterFunction = None
+		self.__filterFunctions = {}
 
 		# Build the layout
 		self.__update()
@@ -226,23 +228,32 @@ class PlugLayout( GafferUI.Widget ) :
 
 		return items
 
-	# Sets the function used to determine which plugs are visible. `filterFunction` is
-	# a function that accepts a `Gaffer.Plug` and returns `True` if the Plug should be
-	# visible and `False` otherwise. This filtering is in addition to metadata-defined
-	# visibility activators. Passing `None` to `filterFunction` disables filtering.
-	def setFilter( self, filterFunction ) :
+	# Adds a function identified by `name` to be used to determine plug visibility.
+	# `filterFunction` is a function that accepts a `Gaffer.Plug` and returns `True`
+	# if the Plug should be visible and `False` otherwise. A plug's visibility is
+	# determined by the intersection of all plug filter functions. This filtering is
+	# in addition to metadata-defined visibility activators.
+	def setFilter( self, name, filterFunction ) :
 
-		if self.__filterFunction == filterFunction :
+		if self.getFilter( name ) == filterFunction :
 			return
-		self.__filterFunction = filterFunction
+		self.__filterFunctions[name] = filterFunction
 		self.__activationsDirty = True
 		self.__updateLazily()
 
 
-	# Returns the function used to determine which plugs are visible.
-	def getFilter( self ) :
+	# Returns the function used to determine plug visibility for a named filter.
+	def getFilter( self, name ) :
 
-		return self.__filterFunction
+		return self.__filterFunctions.get( name, None )
+
+	# Removes a named filter function.
+	def removeFilter( self, name ) :
+
+		if name in self.__filterFunctions :
+			del self.__filterFunctions[name]
+			self.__activationsDirty = True
+			self.__updateLazily()
 
 	@GafferUI.LazyMethod()
 	def __updateLazily( self ) :
@@ -267,6 +278,14 @@ class PlugLayout( GafferUI.Widget ) :
 		# layout from the section definitions.
 
 		self.__layout.update( self.__rootSection )
+
+		# \todo Should this apply to nodes as well?
+		autoexpandThreshold = self.__itemMetadataValue( self.__parent, "autoexpandThreshold" ) or 0 if not isinstance( self.__parent, Gaffer.Node ) else 0
+		if autoexpandThreshold > 0 :
+			visibleItems = [ item for item, widget in self.__widgets.items() if widget is not None and widget.getVisible() ]
+			if len( visibleItems ) <= autoexpandThreshold :
+				visibleSections = set( ".".join( self.__sectionPath( i ) ) for i in visibleItems )
+				self.__layout.expandSections( visibleSections )
 
 	def __updateLayout( self ) :
 
@@ -365,8 +384,8 @@ class PlugLayout( GafferUI.Widget ) :
 			if widget is not None :
 				with self.context() :
 					widget.setEnabled( active( self.__itemMetadataValue( item, "activator" ) ) )
-					visibleByFilter = self.__filterFunction( item ) if ( self.__filterFunction is not None and isinstance( item, Gaffer.Plug ) ) else True
-					widget.setVisible( visibleByFilter and active( self.__itemMetadataValue( item, "visibilityActivator" ) ) )
+					visibleByFilters = all( f( item ) for f in self.__filterFunctions.values() if isinstance( item, Gaffer.Plug ) )
+					widget.setVisible( visibleByFilters and active( self.__itemMetadataValue( item, "visibilityActivator" ) ) )
 
 	def __updateSummariesWalk( self, section ) :
 
@@ -571,6 +590,103 @@ class PlugLayout( GafferUI.Widget ) :
 		self.__summariesDirty = True
 		self.__updateLazily()
 
+# A widget for filtering child plugs of a layout by the label of the plug.
+#
+# Per-widget metadata support :
+#	<layoutName> corresponds to `PlugLayout.__layoutName` for the ancestor `PlugLayout`.
+#	Common layout names include `layout` and `toolbarLayout`.
+#	- "<layoutName>:filterEnabled" controls whether or not the filter will be applied
+#	- "<layoutName>:filter" controls the string used for filtering plugs
+class StandardFilterWidget( GafferUI.Widget ) :
+
+	def __init__( self, parent, **kw ) :
+
+		self.__listContainer = GafferUI.ListContainer( GafferUI.ListContainer.Orientation.Horizontal, spacing = 4 )
+
+		GafferUI.Widget.__init__( self, self.__listContainer, **kw )
+
+		self.__parentComponent = parent
+
+		self.__filterEnabled = False
+
+		with self.__listContainer :
+
+			self.__filterButton = GafferUI.Button( "", "search.png", hasFrame = False )
+			if "toolTip" in kw :
+				self.__filterButton.setToolTip( kw["toolTip"] )
+			self.__filterButton.clickedSignal().connect( Gaffer.WeakMethod( self.__filterButtonClicked ) )
+
+			self.__filterText = GafferUI.TextWidget( "", placeholderText = "Filter..." )
+			self.__filterText.textChangedSignal().connect( Gaffer.WeakMethod( self.__filterTextChanged ) )
+			self.__filterText.editingFinishedSignal().connect( Gaffer.WeakMethod( self.__filterEditingFinished ) )
+
+		self.parentChangedSignal().connect( Gaffer.WeakMethod( self.__parentChanged ) )
+
+		self.__filterEnabled = Gaffer.Metadata.value( self.__parentComponent, "standardFilterWidget:filterEnabled" ) or False
+		self.__filterText.setText( Gaffer.Metadata.value( self.__parentComponent, "standardFilterWidget:filter" ) or "" )
+
+	def __parentChanged( self, widget ) :
+
+		self.__updateWidgets()
+		self.__updateFilter()
+
+	def __plugFilter( self, filter, plug ) :
+
+		label = Gaffer.Metadata.value( plug, "label" ).lower()
+		return IECore.StringAlgo.matchMultiple( label, filter )
+
+	def __plugLayout( self ) :
+
+		return self.ancestor( GafferUI.PlugLayout )
+
+	def __updateFilter( self ) :
+
+		if self.__plugLayout() is None :
+			return
+
+		if self.__filterEnabled :
+			filterValue = self.__filterText.getText().lower()
+			filterValue = filterValue if IECore.StringAlgo.hasWildcards( filterValue ) else ( "*" + filterValue + "*" )
+			self.__plugLayout().setFilter( "standard", functools.partial( Gaffer.WeakMethod( self.__plugFilter ), filterValue ) )
+		else :
+			self.__plugLayout().removeFilter( "standard" )
+
+	def __filterButtonClicked( self, button ) :
+
+		self.__filterEnabled = not self.__filterEnabled
+		Gaffer.Metadata.registerValue( self.__parentComponent, "standardFilterWidget:filterEnabled", self.__filterEnabled, persistent = False )
+
+		if self.__filterEnabled :
+			self.__filterText.setSelection( 0, None )  # All
+			self.__filterText.grabFocus()
+
+		self.__updateFilter()
+		self.__updateWidgets()
+
+	def __filterTextChanged( self, textWidget ) :
+
+		self.__updateFilter()
+
+	def __filterEditingFinished( self, textWidget ) :
+
+		Gaffer.Metadata.registerValue( self.__parentComponent, "standardFilterWidget:filter", textWidget.getText(), persistent = False )
+
+	def __updateWidgets( self ) :
+
+		self.__filterButton.setImage( "searchOn.png" if self.__filterEnabled else "search.png" )
+		self.__filterText.setVisible( self.__filterEnabled )
+
+
+PlugLayout.StandardFilterWidget = StandardFilterWidget
+
+class PlugLabelSpacer( GafferUI.Spacer ) :
+
+	def __init__( self, parent, **kw ) :
+
+		s = imath.V2i( GafferUI.PlugWidget.labelWidth() + 9, 18 )
+		GafferUI.Spacer.__init__( self, s, maximumSize = s )
+
+PlugLayout.StandardFilterWidget.PlugLabelSpacer = PlugLabelSpacer
 
 class _AccessoryRow( GafferUI.ListContainer ) :
 
@@ -643,6 +759,10 @@ class _Layout( GafferUI.Widget ) :
 	# Returns `True` if the layout contains any visible widgets,
 	# `False` otherwise.
 	def update( self, section ) :
+
+		raise NotImplementedError
+
+	def expandSections( self, sections ) :
 
 		raise NotImplementedError
 
@@ -722,6 +842,10 @@ class _TabLayout( _Layout ) :
 
 		return haveVisibleWidgets or haveVisibleTabs
 
+	def expandSections( self, sections ) :
+		# \todo If len( sections ) == 1 expand it, otherwise... pick the first or last? Don't do anything?
+		pass
+
 	def __currentTabChanged( self, tabbedContainer, currentTab ) :
 
 		self.__section.saveState( "currentTab", tabbedContainer.index( currentTab ) )
@@ -735,6 +859,7 @@ class _CollapsibleLayout( _Layout ) :
 		_Layout.__init__( self, self.__column, orientation, **kw )
 
 		self.__collapsibles = {} # Indexed by section name
+		self.__collapsibleStateChangedSignals = {}
 
 	def update( self, section ) :
 
@@ -759,14 +884,15 @@ class _CollapsibleLayout( _Layout ) :
 				# way of controlling size behaviours for all widgets in the public API.
 				collapsible.getCornerWidget()._qtWidget().setSizePolicy( QtWidgets.QSizePolicy.Ignored, QtWidgets.QSizePolicy.Fixed )
 
-				if subsection.restoreState( "collapsed" ) is False :
-					collapsible.setCollapsed( False )
-
-				collapsible.stateChangedSignal().connect(
-					functools.partial( Gaffer.WeakMethod( self.__collapsibleStateChanged ), subsection = subsection )
+				self.__collapsibleStateChangedSignals[name] = collapsible.stateChangedSignal().connect(
+					functools.partial( Gaffer.WeakMethod( self.__collapsibleStateChanged ), subsection = subsection ),
+					scoped = True
 				)
 
 				self.__collapsibles[name] = collapsible
+
+			with Gaffer.Signals.BlockedConnection( self.__collapsibleStateChangedSignals[name] ) :
+				collapsible.setCollapsed( False if subsection.restoreState( "collapsed" ) is False else True )
 
 			collapsible.setVisible(
 				collapsible.getChild().update( subsection )
@@ -787,9 +913,27 @@ class _CollapsibleLayout( _Layout ) :
 
 		return any( w.getVisible() for w in self.__column )
 
+	def expandSections( self, sections ) :
+
+		if len( sections ) == 0 :
+			return
+
+		for s in sections :
+			self.__expandSectionWalk( s )
+
 	def __collapsibleStateChanged( self, collapsible, subsection ) :
 
 		subsection.saveState( "collapsed", collapsible.getCollapsed() )
+
+	def __expandSectionWalk( self, section ) :
+
+		section, delimiter, childSection = section.partition( "." )
+		if childSection != "" :
+			self.__expandSectionWalk( childSection )
+
+		if section in self.__collapsibles :
+			with Gaffer.Signals.BlockedConnection( self.__collapsibleStateChangedSignals[section] ) :
+				self.__collapsibles[section].setCollapsed( False )
 
 class _MissingCustomWidget( GafferUI.Widget ) :